requirements.txt
Flask==2.0.2
flask-restx==0.5.1
Flask-Cors==3.0.10
gunicorn==20.1.0
python-dotenv==0.19.2
langdetect==1.0.9
requests==2.26.0
Werkzeug==2.0.3
